+++ resources/views/individuelles/show.blade.php
@@ -15,50 +15,50 @@
     <div class="container">
         <div class="container-fluid">
             <div class="card">
-                    <div class="card-body">
-                        <h1 class="h4 h-100 text-uppercase mb-4"><b><u>Type de demande</u> : {!! __('Individuelle') !!}
-                            </b></h1>
-                        <h4>
-                            <b><u>N° du dossier</u></b> : <span
-                                class="font-italic">{{ $individuelle->demandeur->numero ?? 'numéro de dossier introuvable' }}</span>
-                        </h4><br />
-                        <h4>
-                            <?php $i = 1; ?>
-                            <b><u>Modules demandés </u></b>: 
-                            @foreach ($individuelle->modules as $module)
-                              <small>{!! $i++ !!}</small>)
-                                    {!! $module->name ?? 'aucun module demandé' !!}</small>
-                            @endforeach
-                        </h4><br />
-                        <h4>
-                            @if (isset($individuelle->demandeur->programme->sigle))
-                                <b><u>Programmes </u></b>:
-                                {!! ucwords(strtolower($individuelle->demandeur->programme->sigle)) ?? 'aucun module demandé' !!}</small>
-                            @else
+                <div class="card-body">
+                    <h1 class="h4 h-100 text-uppercase mb-4"><b><u>Type de demande</u> : {!! __('Individuelle') !!}
+                        </b></h1>
+                    <h4>
+                        <b><u>N° du dossier</u></b> : <span
+                            class="font-italic">{{ $individuelle->demandeur->numero ?? 'numéro de dossier introuvable' }}</span>
+                    </h4><br />
+                    <h4>
+                        <?php $i = 1; ?>
+                        <b><u>Modules demandés </u></b>:
+                        @foreach ($individuelle->modules as $module)
+                            <small>{!! $i++ !!}</small>)
+                            {!! $module->name ?? 'aucun module demandé' !!}</small>
+                        @endforeach
+                    </h4><br />
+                    <h4>
+                        @if (isset($individuelle->demandeur->programme->sigle))
+                            <b><u>Programmes </u></b>:
+                            {!! ucwords(strtolower($individuelle->demandeur->programme->sigle)) ?? 'aucun module demandé' !!}</small>
+                        @else
 
-                            @endif
-                        </h4><br /><br />
-                        <div class="d-flex justify-content-between align-items-center mt-3">
-                            <a href="{!! url('individuelles/' . $individuelle->id . '/edit') !!}" title="modifier" class="btn btn-outline-warning">
-                                <i class="far fa-edit">&nbsp;Modifier</i>
-                            </a>
-                            <a href="{!! url('individuelles/' . $individuelle->demandeur->id . '/edit') !!}" title="voir les d&eacute;tails du courrier"
-                                class="btn btn-outline-primary">
-                                <i class="far fa-eye">&nbsp;D&eacute;tails</i>
-                            </a>
-                                {!! Form::open(['method' => 'DELETE', 'url' => 'individuelles/' . $individuelle->id, 'id' => 'deleteForm', 'onsubmit' => 'return ConfirmDelete()']) !!}
-                                {!! Form::button('<i class="fa fa-trash">&nbsp;Supprimer</i>', ['type' => 'submit', 'class' => 'btn btn-outline-danger', 'title' => 'supprimer']) !!}
-                                {!! Form::close() !!}
-                        </div>
-                        <div class="d-flex justify-content-between align-items-center mt-5">
-                            <small>Déposée depuis le {!! Carbon\Carbon::parse($individuelle->demandeur->created_at)->format('d/m/Y à H:i:s') !!}</small>
-                            <span class="badge badge-primary">{!! $individuelle->demandeur->user->firstname !!}&nbsp;{!! $individuelle->demandeur->user->name !!}</span>
-                        </div>
+                        @endif
+                    </h4><br /><br />
+                    <div class="d-flex justify-content-between align-items-center mt-3">
+                        <a href="{!! url('individuelles/' . $individuelle->id . '/edit') !!}" title="modifier" class="btn btn-outline-warning">
+                            <i class="far fa-edit">&nbsp;Modifier</i>
+                        </a>
+                        <a href="{!! url('individuelles/' . $individuelle->demandeur->id . '/edit') !!}" title="voir les d&eacute;tails du courrier"
+                            class="btn btn-outline-primary">
+                            <i class="far fa-eye">&nbsp;D&eacute;tails</i>
+                        </a>
+                        @can('user-delete')
+                            {!! Form::open(['method' => 'DELETE', 'url' => 'individuelles/' . $individuelle->id, 'id' => 'deleteForm', 'onsubmit' => 'return ConfirmDelete()']) !!}
+                            {!! Form::button('<i class="fa fa-trash">&nbsp;Supprimer</i>', ['type' => 'submit', 'class' => 'btn btn-outline-danger', 'title' => 'supprimer']) !!}
+                            {!! Form::close() !!}
+                        @endcan
                     </div>
+                    <div class="d-flex justify-content-between align-items-center mt-5">
+                        <small>Déposée depuis le {!! Carbon\Carbon::parse($individuelle->demandeur->created_at)->format('d/m/Y à H:i:s') !!}</small>
+                        <span class="badge badge-primary">{!! $individuelle->demandeur->user->firstname !!}&nbsp;{!! $individuelle->demandeur->user->name !!}</span>
+                    </div>
+                </div>
             </div>
-
             <hr>
-
             {{-- <div class="card">
                     <div class="card-body">
                         <h3 class="card-title text-center">Commentaires</h5>
